+++ src/test/java/selenium/steps/TheShopSteps.java
@@ -1,26 +1,25 @@
-package selenium;
-
-import io.cucumber.spring.CucumberContextConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import com.example.produktapi.ProduktapiApplication;
+package selenium.steps;
 
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+import io.cucumber.spring.CucumberContextConfiguration;
+import org.junit.jupiter.api.Assertions;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import com.example.produktapi.ProduktapiApplication;
 import io.github.bonigarcia.wdm.WebDriverManager;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-@SpringBootTest(classes = ProduktapiApplication.class)
 @CucumberContextConfiguration
+@SpringBootTest(classes = ProduktapiApplication.class)
 public class TheShopSteps {
 
-    WebDriver driver;
-    String URL = "https://webshop-agil-testautomatiserare.netlify.app/";
+    private WebDriver driver;
+    private String URL = "https://webshop-agil-testautomatiserare.netlify.app/";
 
     @Given("I open the web shop")
     public void iOpenTheWebShop() {
@@ -35,16 +34,27 @@ public class TheShopSteps {
         driver.get(URL);
     }
 
+    @When("I check the page title")
+    public void iCheckThePageTitle() {
+        // This step can be empty or perform some action
+    }
+
     @Then("the title should be {string}")
     public void theTitleShouldBe(String expectedTitle) {
-        assertEquals(expectedTitle, driver.getTitle());
+        String actualTitle = driver.getTitle();
+        Assertions.assertEquals(expectedTitle, actualTitle, "Page title does not match!");
         driver.quit();
+    }
+
+    @When("I look for the {string} button")
+    public void iLookForTheButton(String buttonText) {
+        // This step can be empty or perform some action
     }
 
     @Then("the {string} button should be visible")
     public void theButtonShouldBeVisible(String buttonText) {
         String actualText = driver.findElement(By.xpath("/html/body/div[1]/div/div[1]/div/button")).getText();
-        assertEquals(buttonText, actualText);
+        Assertions.assertEquals(buttonText, actualText, "Button text does not match!");
         driver.quit();
     }
 }
